working on the client

+++ server/src/main/java/com/workmate/server/repository/CompanyRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository("companyRepository")
-public interface CompanyRepository extends JpaRepository<Company, Integer>
+public interface CompanyRepository extends JpaRepository<Company, Long>
 {
     Company findByName(String name);
 }
